Extract sandbox to separate executable as testcase viewer
