+++ src/app/CMakeLists.txt
@@ -1,4 +1,4 @@
-target_sources(achtung_app PRIVATE
+target_sources(freeachtung_app PRIVATE
     AssetManager.cpp
     Application.cpp
     ConfigManager.cpp
