+++ src/engine/CMakeLists.txt
@@ -1,8 +1,6 @@
-target_sources(achtung_app PRIVATE
+target_sources(freeachtung_app PRIVATE
     BorderThing.cpp
     Engine.cpp
-    EngineTestable.cpp
-    InputSequence.cpp
     Player.cpp
     PickMeUp.cpp
     TimerService.cpp
