+++ test/testable/InputSequence.hpp
