+++ test/testcase_preview/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(testcase_preview PUBLIC
+    Main.cpp
+    StateViewer.cpp
+)
+
+target_link_libraries(testcase_preview PUBLIC freeachtung_app testable)
